// src/lib/models/api/PagedResponse.ts
import { Pageable } from './Pageable';
import { Sort } from './Sort';

export interface PagedResponse<T> {
    content: T[];
    pageable: Pageable;
    last: boolean;
    totalPages: number;
    totalElements: number;
    size: number;
    number: number;
    sort: Sort;
    first: boolean;
    numberOfElements: number;
    empty: boolean;
}